+++ examples/General/SettingUserNotifications.php
@@ -40,7 +40,6 @@ try {
 
     $ig->event->sendNavigation('main_search', 'feed_timeline', 'explore_popular', null, null, $topicData);
     $ig->discover->getNullStateDynamicSections();
-    $ig->discover->getSuggestedSearches('blended');
     $sectionalItems = $ig->discover->getExploreFeed('explore_all:0', $searchSession)->getSectionalItems();
 
     $ig->event->prepareAndSendExploreImpression('explore_all:0', $searchSession, $sectionalItems);