+++ examples/Login/twoFactorLogin.php
@@ -6,8 +6,8 @@ date_default_timezone_set('UTC');
 require __DIR__.'/../../vendor/autoload.php';
 
 /////// CONFIG ///////
-$username = 'hellogloria_';
-$password = 'OldH4gs!';
+$username = '';
+$password = '';
 $debug = true;
 $truncatedDebug = false;
 //////////////////////
@@ -20,24 +20,15 @@ try {
     if ($loginResponse !== null && $loginResponse->isTwoFactorRequired()) {
         $twoFactorInfo = $loginResponse->getTwoFactorInfo();
         $twoFactorIdentifier = $twoFactorInfo->getTwoFactorIdentifier();
-
-        if ($twoFactorInfo->getTotpTwoFactorOn()) {
-            $method = 3; // Two Factor Application
-            $showMessage = 'Enter the code in your TOTP application';
-        } elseif ($twoFactorInfo->getWhatsappTwoFactorOn()) {
-            $method = 6; // Code via Whatsapp
-            $showMessage = 'Enter the code you received via WhatsApp';
-        } elseif ($twoFactorInfo->getSmsTwoFactorOn()) {
-            $method = 1; // Code via SMS
-            $showMessage = 'Enter the code you received via SMS';
+        if (($twoFactorInfo->getSmsTwoFactorOn() === true) && ($twoFactorInfo->getTotpTwoFactorOn() === false)) {
+            $method = 1; // SMS
         } else {
-            throw new Exception('Unknown Two Factor method');
+            $method = 3; // TOTP
         }
 
         // The "STDIN" lets you paste the code via terminal for testing.
         // You should replace this line with the logic you want.
         // The verification code will be sent by Instagram via SMS.
-        echo "{$showMessage}: ";
         $verificationCode = trim(fgets(STDIN));
         $ig->finishTwoFactorLogin($username, $password, $twoFactorIdentifier, $verificationCode, $method);
     }